+++ apps/lib/build.info
@@ -9,7 +9,7 @@ ENDIF
 
 # Source for libapps
 $LIBAPPSSRC=apps.c apps_ui.c opt.c fmt.c s_cb.c s_socket.c app_rand.c \
-        columns.c app_params.c names.c app_provider.c app_x509.c
+        columns.c app_params.c names.c app_provider.c app_x509.c http_server.c
 
 IF[{- !$disabled{apps} -}]
   LIBS{noinst}=../libapps.a
